pyproject.toml
[tool.poetry]
name = "flask-test"
version = "0.1.0"
description = ""
authors = ["Petr <p>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask_cors = "^3.0.8"
flask_restplus = "^0.13.0"
flask_sqlalchemy = "^2.4.4"
rq = "^1.5.0"
requests = "^2.24.0"
Werkzeug = "0.16.0"
loguru = "^0.5.1"
alembic = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
